+++ src/frontend/apps/standalone_site/src/features/Contents/features/ClassRoom/components/Update/ClassRoomUpdate.tsx
@@ -9,7 +9,6 @@ import {
   isDecodedJwtWeb,
   useCurrentResourceContext,
   useJwt,
-  localStore as useLocalJwt,
   useResponsive,
 } from 'lib-components';
 import { useEffect, useMemo } from 'react';
@@ -70,20 +69,9 @@ const DashboardClassroomStyled = styled(Box)<DashboardClassroomStyledProps>`
 `;
 
 const ClassRoomUpdate = () => {
-  const { classroomId, inviteId } = useParams();
+  const { classroomId } = useParams();
 
-  const internalDecodedPersistingJwt = useJwt(
-    (state) => state.internalDecodedJwt,
-  );
-  const internalDecodedLocalJwt = useLocalJwt(
-    (state) => state.internalDecodedJwt,
-  );
-
-  // Invited users have a local jwt (working only for the current browser tab),
-  // otherwise it's a persisting jwt
-  const internalDecodedJwt = inviteId
-    ? internalDecodedLocalJwt
-    : internalDecodedPersistingJwt;
+  const internalDecodedJwt = useJwt((state) => state.internalDecodedJwt);
   const resourceContext: ResourceContext = useMemo(() => {
     let permissions = {
       can_access_dashboard: false,